Combo-box update
Update the combo box to create two different boxes, one for functions and the other for commands. Add a button for sending or connecting. Another update is for the test file, which is responsible for loading the build from the database and creating the main GUI, and will be replaced by Shira.

Co-Authored-By: ShiraGit1 <57354875+ShiraGit1@users.noreply.github.com>